requirements.txt
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.4
django-crispy-forms==1.9.2
django-gst-field==0.0.3
django-paypal==1.0.0
django-phone-field==1.8.1
django-phonenumber-field==4.0.0
gunicorn==20.0.4
idna==2.10
phonenumbers==8.12.6
Pillow==6.2.1
pytz==2018.7
requests==2.24.0
six==1.15.0
urllib3==1.25.9
whitenoise==5.1.0
psycopg2==2.7.2